tx2: use in-memory DTB coming from cboot
CBoot loads the DTB from a partition and modifies it dynamically. We
need those modifications so we get persistent MAC addresses, USB works,
and to prevent kernel crashes. So, we do not load the DTB anymore from
the system partition and instead we use the one already loaded and
modified by CBoot. The problem that we have now is that the DTB won't be
updated when the kernel snap gets updated, so we will have to revisit
this.