nano: add content of cbootargs to our kernel command line
The arguments cboot creates change depending on whether we have a
connected HDMI monitor or not, and if we do not use the right ones the
kernel crashes when we do have a monitor connected. Make use of them
then. Interestingly, cbootargs shows the same behaviour in the TX1 (same
SoC), but in that case the kernel does not crash...